capi: move all C-specific stuff into capi
This will allow us to mark the C API stuff as #[cfg(capi)] to let us
build Rust crates without any C code.

While we're at it, split out the CError stuff from the capi::ret and
capi::utils modules to make it easier to see where things fit.
